bdm.mpi.stop() bug

+++ package/src/tsne.cpp
@@ -29,11 +29,11 @@
 using namespace std;
 
 // t-SNE constructor
-TSNE::TSNE(size_t z, size_t w, size_t mY, double* eRange, int max_iter, double lRate, double theta, double alpha, double zP, double* exgg) : z(z), w(w), mY(mY), eRange(eRange), max_iter(max_iter), lRate(lRate), theta(theta), alpha(alpha), zP(zP), exgg(exgg)
+TSNE::TSNE(unsigned int z, unsigned int w, unsigned int mY, double* eRange, int max_iter, double lRate, double theta, double alpha, double zP, double* exgg) : z(z), w(w), mY(mY), eRange(eRange), max_iter(max_iter), lRate(lRate), theta(theta), alpha(alpha), zP(zP), exgg(exgg)
 {
 	// set current value for learning rate
 	std::vector<double> eta (mY, 0);
-	for (size_t d = 0; d < mY; d++) eta[d] = lRate *4.0;
+	for (unsigned int d = 0; d < mY; d++) eta[d] = lRate *4.0;
 	this ->eta = eta;
 }
 
@@ -55,8 +55,8 @@ void TSNE::run2D(double* P, int* W, double* Y)
 			zQ = apprx_Gradient(P, W, Y, atrF, repF);
 		}
 		// update (with momentum and learning rate)
-		for (size_t i = 0, k = 0; i < z; i++){
-			for (size_t d = 0; d < mY; d++, k++){
+		for (unsigned int i = 0, k = 0; i < z; i++){
+			for (unsigned int d = 0; d < mY; d++, k++){
 				// update embedding position
 				uY[k] = alpha *uY[k] -eta[d] *(exgg[i] *atrF[k] /zP -repF[k] /zQ);
 				Y[k] += uY[k];
@@ -69,8 +69,8 @@ void TSNE::run2D(double* P, int* W, double* Y)
 	// eRange is used to compute the embedding size (i.e. pooled from the master)
 	// if not used to compute the learning-rate
 	// there is no need to push it from the master (remove pushing in bdm_ptsne.R !!)
-	for (size_t i = 0, k = 0; i < z; i++){
-		for (size_t d = 0; d < mY; d++, k++){
+	for (unsigned int i = 0, k = 0; i < z; i++){
+		for (unsigned int d = 0; d < mY; d++, k++){
 			if (Y[k] < eRange[d *mY +0]) eRange[d *mY +0] = Y[k];
 			else if (Y[k] > eRange[d *mY +1]) eRange[d *mY +1] = Y[k];
 		}
@@ -87,15 +87,15 @@ double TSNE::exact_Gradient(double* P, double* Y, double* atrF, double* repF)
 {
 	double zQ = .0;
 	double L [mY];
-	for(size_t i = 0, ij = 0; i < z; i++) {
-		for (size_t j = i +1; j < z; j++, ij++) {
+	for(unsigned int i = 0, ij = 0; i < z; i++) {
+		for (unsigned int j = i +1; j < z; j++, ij++) {
 			double Lij = 1.0;
-			for(size_t d = 0; d < mY; d++) {
+			for(unsigned int d = 0; d < mY; d++) {
 				L[d] = Y[i *mY +d] -Y[j *mY +d];
 				Lij += L[d] *L[d];
 			}
 			double Qij = 1.0 /Lij;
-			for(size_t d = 0; d < mY; d++) {
+			for(unsigned int d = 0; d < mY; d++) {
 				if (P[ij] > 0) {
 					double Aij = P[ij] *Qij *L[d];
 					atrF[i *mY +d] += Aij;
@@ -117,16 +117,16 @@ double TSNE::apprx_Gradient(double* P, int* W, double* Y, double* atrF, double* 
 {
 	// compute attractive forces
 	double L [mY];
-	for(size_t k = 0; k < w; k++) {
-		size_t i = W[k] /z;
-		size_t j = (int) W[k] %z;
+	for(unsigned int k = 0; k < w; k++) {
+		unsigned int i = W[k] /z;
+		unsigned int j = (int) W[k] %z;
 		double Lij = 1.0;
-		for(size_t d = 0; d < mY; d++){
+		for(unsigned int d = 0; d < mY; d++){
 			L[d] = Y[i *mY +d] - Y[j *mY +d];
 			Lij +=  L[d] *L[d];
 		}
-		size_t ij = ijIdx(z, i, j);
-		for(size_t d = 0; d < mY; d++) {
+		unsigned int ij = ijIdx(z, i, j);
+		for(unsigned int d = 0; d < mY; d++) {
 			double Aij = P[ij] *L[d] /Lij;
 			atrF[i *mY +d] += Aij;
 			atrF[j *mY +d] -= Aij;
@@ -150,10 +150,10 @@ double TSNE::getCost(double* P, double* Y)
 	double zQ = .0;
 	// join cross entropy
 	double jxH = .0;
-	for (size_t i = 0, ij = 0; i < z; i++) {
-		for (size_t j = i +1; j < z; j++, ij++){
+	for (unsigned int i = 0, ij = 0; i < z; i++) {
+		for (unsigned int j = i +1; j < z; j++, ij++){
 			double Lij = 1.0;
-			for(size_t d = 0; d < mY; d++) Lij += std::pow(Y[i *mY +d] -Y[j *mY +d], 2);
+			for(unsigned int d = 0; d < mY; d++) Lij += std::pow(Y[i *mY +d] -Y[j *mY +d], 2);
 			if (P[ij] > 0) {
 				jxH -= P[ij] *std::log(Lij);
 			}
